// sourcecode/BTrak.Packages/snova-role-management/role-module-components/role-components/src/lib/snova-role-management/models/entity-type-role-feature-model.ts
export class EntityTypeRoleFeatureModel {
    entityFeatureIds: string[];
    entityRoleId: string;
    isArchived: boolean;
    entityRoleFeatureId: string;
    entityFeatureId: string;
    entityFeatureName: string;
    entityRoleName: string;
    projectId: string;
    createdDateTime: Date;
    inActiveDateTime: Date;
    totalCount: number;
    timeStamp: any;
}