// sourcecode/BTrak.Packages/snova-role-management/role-module-components/role-components/src/lib/snova-role-management/models/entity-role-model.ts
export class EntityRoleModel {
    entityRoleId: string;
    entityRoleName: string;
    companyId: string;
    createdDateTime: Date;
    createdByUserId: string;
    versionNumber: number;
    inActiveDateTime: string;
    originalId: string;
    asAtInActiveDateTime: Date;
    timeStamp: any;
    isArchived: boolean;
}